HomeWork for lesson02 V-23

+++ src/main/java/com/company/Calculator.java
@@ -18,6 +18,7 @@ public class Calculator {
     }
 
     public double division(int a, int b) {
-        return (a/b);
+
+        return (double) a/b;
     }
 }
